python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stockbuddy"
version = "0.1.0"
description = "StockBuddy is a community-driven, multi-agent platform for financial applications."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-okx>=0.4.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "a2a-sdk[http-server]>=0.3.4",
    "yfinance>=0.2.65",
    "requests>=2.32.5",
    "akshare>=1.17.44",
    "agno[openai, google, lancedb]>=2.0,<3.0",
    "edgartools>=4.12.2",
    "sqlalchemy>=2.0.43",
    "aiosqlite>=0.19.0",
    "unstructured>=0.18.15",
    "markdown>=3.9",
    "loguru>=0.7.3",
    "aiofiles>=24.1.0",
    "crawl4ai>=0.7.4",
    "ccxt>=4.5.15",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.0.0",
    "diff-cover>=9.0.0",
]

[dependency-groups]
dev = [
    {include-group = "lint"},
    {include-group = "style"},
    {include-group = "test"}
]
lint = [
    "ruff"
]
style = [
    "ruff",
    "isort",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.0.0",
    "diff-cover>=9.0.0",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
force-exclude = true
exclude = [
    "venv",
    "build",
    "dist",
    "node_modules",
    "site-packages",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.hatch.build.targets.wheel]
include = [
    "stockbuddy",
]
exclude = [
    "tests/**",
    "**/tests/**",
    "examples/**",
    "**/examples/**",
    "docs/**",
    "**/docs/**",
]

[tool.pytest.ini_options]
norecursedirs = [
    ".venv",
    "venv",
    "build",
    "dist",
    "logs",
    "examples",
    "docs",
]

[tool.isort]
profile = "black"
skip_glob = [
    "tests/**",
    "**/tests/**",
    "docs/**",
    "**/docs/**",
]
